refactor: use dependency for snapshot load type
