+++ backend/qualibrate_app/api/routes/root.py
@@ -1,5 +1,5 @@
 from collections.abc import Sequence
-from typing import Annotated, Any, Optional, Union
+from typing import Annotated, Any, Union
 
 from fastapi import APIRouter, Depends, Query
 from qualibrate_config.models import QualibrateConfig, StorageType
@@ -8,9 +8,7 @@ from qualibrate_app.api.core.domain.bases.branch import BranchLoadType
 from qualibrate_app.api.core.domain.bases.node import NodeLoadType
 from qualibrate_app.api.core.domain.bases.root import RootBase
 from qualibrate_app.api.core.domain.bases.snapshot import (
-    SnapshotLoadType,
     SnapshotLoadTypeFlag,
-    SnapshotLoadTypeToLoadTypeFlag,
 )
 from qualibrate_app.api.core.domain.local_storage.root import RootLocalStorage
 from qualibrate_app.api.core.domain.timeline_db.root import RootTimelineDb
@@ -23,8 +21,8 @@ from qualibrate_app.api.core.models.snapshot import (
 from qualibrate_app.api.core.models.snapshot import Snapshot as SnapshotModel
 from qualibrate_app.api.core.types import IdType
 from qualibrate_app.api.dependencies.search import get_search_path
-from qualibrate_app.api.routes.utils.snapshot_load_type import (
-    parse_load_type_flag,
+from qualibrate_app.api.routes.utils.dependencies import (
+    get_snapshot_load_type_flag,
 )
 from qualibrate_app.config import (
     get_settings,
@@ -82,17 +80,12 @@ def get_latest_node(
 def get_snapshot_by_id(
     *,
     id: IdType,
-    load_type: Annotated[
-        Optional[SnapshotLoadType], Query(deprecated="use load_type_flag")
-    ] = None,
     load_type_flag: Annotated[
-        SnapshotLoadTypeFlag, Depends(parse_load_type_flag)
-    ] = SnapshotLoadTypeFlag.Metadata,
+        SnapshotLoadTypeFlag, Depends(get_snapshot_load_type_flag)
+    ],
     root: Annotated[RootBase, Depends(_get_root_instance)],
 ) -> SnapshotModel:
     snapshot = root.get_snapshot(id)
-    if load_type is not None:
-        load_type_flag = SnapshotLoadTypeToLoadTypeFlag[load_type]
     snapshot.load_from_flag(load_type_flag)
     return snapshot.dump()
 
@@ -100,17 +93,12 @@ def get_snapshot_by_id(
 @root_router.get("/snapshot/latest")
 def get_latest_snapshot(
     *,
-    load_type: Annotated[
-        Optional[SnapshotLoadType], Query(deprecated=True)
-    ] = None,
     load_type_flag: Annotated[
-        SnapshotLoadTypeFlag, Depends(parse_load_type_flag)
+        SnapshotLoadTypeFlag, Depends(get_snapshot_load_type_flag)
     ],
     root: Annotated[RootBase, Depends(_get_root_instance)],
 ) -> SnapshotModel:
     snapshot = root.get_snapshot()
-    if load_type is not None:
-        load_type_flag = SnapshotLoadTypeToLoadTypeFlag[load_type]
     snapshot.load_from_flag(load_type_flag)
     return snapshot.dump()
