+++ backend/qualibrate_app/api/routes/snapshot.py
@@ -9,9 +9,7 @@ from qualibrate_config.models import QualibrateConfig, StorageType
 
 from qualibrate_app.api.core.domain.bases.snapshot import (
     SnapshotBase,
-    SnapshotLoadType,
     SnapshotLoadTypeFlag,
-    SnapshotLoadTypeToLoadTypeFlag,
 )
 from qualibrate_app.api.core.domain.local_storage.snapshot import (
     SnapshotLocalStorage,
@@ -32,8 +30,8 @@ from qualibrate_app.api.core.types import DocumentSequenceType, IdType
 from qualibrate_app.api.core.utils.request_utils import get_runner_config
 from qualibrate_app.api.core.utils.types_parsing import types_conversion
 from qualibrate_app.api.dependencies.search import get_search_path
-from qualibrate_app.api.routes.utils.snapshot_load_type import (
-    parse_load_type_flag,
+from qualibrate_app.api.routes.utils.dependencies import (
+    get_snapshot_load_type_flag,
 )
 from qualibrate_app.config import (
     get_settings,
@@ -56,16 +54,11 @@ def _get_snapshot_instance(
 @snapshot_router.get("/")
 def get(
     *,
-    load_type: Annotated[
-        Optional[SnapshotLoadType], Query(deprecated=True)
-    ] = None,
     load_type_flag: Annotated[
-        SnapshotLoadTypeFlag, Depends(parse_load_type_flag)
+        SnapshotLoadTypeFlag, Depends(get_snapshot_load_type_flag)
     ],
     snapshot: Annotated[SnapshotBase, Depends(_get_snapshot_instance)],
 ) -> SnapshotModel:
-    if load_type is not None:
-        load_type_flag = SnapshotLoadTypeToLoadTypeFlag[load_type]
     snapshot.load_from_flag(load_type_flag)
     return snapshot.dump()
